+++ packages/react/src/atomic/List/src/list.stories.tsx
@@ -39,3 +39,9 @@ export const ordered: Story = {
     as: "ol",
   },
 };
+
+export const customPadding: Story = {
+  args: {
+    padding: "2",
+  },
+};