+++ packages/react/src/composite/Collapsible/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The Collapsible is a component that can hid and show another component when a event is triggered from a button or a link.
 
+## 2026-01-13 `1.1.1`
+
+#### 🎉 New features
+
+- Adds support for React 19. ([#404](https://github.com/TiendaNube/nimbus-design-system/pull/404) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-03-18 `1.1.0`
 
 #### 🎉 New features