+++ packages/react/src/composite/ScrollPane/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 The ScrollPane component handles responsive scrolls in lists that are overflowing inside a container.
 
+## 2026-01-13 `1.1.3`
+
+#### 🎉 New features
+
+- Adds support for React 19. ([#404](https://github.com/TiendaNube/nimbus-design-system/pull/404) by [@joacotornello](https://github.com/joacotornello))
+
+## 2026-01-13 `1.1.2`
+
+#### 🐛 Bug fixes
+
+- `ScrollPane`: Fixed types to avoid bloating the documentation builder. ([#350](https://github.com/TiendaNube/nimbus-design-system/pull/350) by [@joacotornello](https://github.com/joacotornello))
+
+## 2025-12-01 `1.1.1`
+
+#### 🐛 Bug fixes
+
+- `ScrollPane`: Fixed focus ring styles to match design system. ([#383](https://github.com/TiendaNube/nimbus-design-system/pull/383) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-11-12 `1.1.0`
 
 #### 🎉 New features